+++ Banking/src/main/java/com/lcf/app/services/AddressService.java
@@ -4,6 +4,6 @@ import com.lcf.app.beans.AddressDetails;
 
 public interface AddressService {
 
-	public abstract int insertAddress(AddressDetails address);
-	public abstract AddressDetails getAddressById(int addressId);
+	public abstract long insertAddress(AddressDetails address);
+	public abstract AddressDetails getAddressById(long addressId);
 }
